Refactor offset resolvers into separate module
