+++ src/__tests__/buildSchema.spec.ts
@@ -9,7 +9,7 @@ import { GraphQLBoolean, GraphQLString, GraphQLFloat,
 import { GraphQLDateTime, GraphQLDate, GraphQLTime } from 'graphql-iso-date'
 import { makeObjects, buildGraphQLObjects } from '../buildSchema'
 import { joinNames } from '../util'
-import { middleware as offsetMiddleware } from '../Pagination/Offset'
+import { middleware as offsetMiddleware } from '../Pagination/Offset/Offset'
 // tslint:disable:no-expression-statement
 
 const objects: Array<[DescribeSObjectResult, SalesforceObjectConfig | undefined]>
